requirements.txt
# メイン依存関係（プロジェクトルート）
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1

# 非同期データベース
aiosqlite==0.19.0
asyncpg==0.29.0
redis[hiredis]==5.0.1

# JSON処理（高速・安全）
orjson==3.9.10

# 認証・セキュリティ（OWASP ASVS Level 2準拠）
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7  # 高度な暗号化機能
bcrypt==4.1.2  # パスワードハッシュ化
python-multipart==0.0.6
slowapi==0.1.9

# PaSoRi関連
nfcpy==1.0.4
# または felica-pasori（どちらか選択）
# felica-pasori==0.1.0

# 開発・テスト用
flake8==6.1.0
mypy==1.7.1
black==23.11.0
requests==2.31.0
pytest-playwright==0.4.3
playwright==1.40.0
pytest-html==4.1.1

# 通知・その他
slack-sdk==3.26.1
schedule==1.2.1
pandas==1.5.3
aiohttp==3.9.1

# レポート・エクスポート関連
openpyxl==3.1.2
reportlab==4.0.8

# データ分析・統計
numpy==1.24.4
statistics==1.0.3.5
